Experiment

+++ Banking/src/main/java/com/lcf/app/services/CustomerService.java
@@ -4,7 +4,7 @@ import com.lcf.app.beans.CustomerDetails;
 
 public interface CustomerService {
 	
-	public abstract int newCustomer(CustomerDetails customer);
+	public abstract long newCustomer(CustomerDetails customer);
 	public abstract CustomerDetails getCustomerById(long id);
 
 }
